refactor: move objects to separate files
Move `ModuleStrictGenerateCommon` and
`StrictGenerateProps` to separate files.
